requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==2.2.3
connexion[swagger-ui]==2.14.2